# ENH: Reduce torch transfers with a torch